+++ src/Symfony/Component/DependencyInjection/Tests/Dumper/YamlDumperTest.php
@@ -80,6 +80,13 @@ class YamlDumperTest extends TestCase
         $this->assertStringEqualsFile(self::$fixturesPath.'/yaml/services24.yml', $dumper->dump());
     }
 
+    public function testDumpAutowireWithOptionalParametersDisabledData()
+    {
+        $container = include self::$fixturesPath.'/containers/container25.php';
+        $dumper = new YamlDumper($container);
+        $this->assertStringEqualsFile(self::$fixturesPath.'/yaml/services25.yml', $dumper->dump());
+    }
+
     public function testDumpDecoratedServices()
     {
         $container = include self::$fixturesPath.'/containers/container34.php';