+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/containers/container25.php
@@ -1,0 +1,14 @@
+<?php
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+
+$container = new ContainerBuilder();
+
+$container
+    ->register('foo', 'Foo')
+    ->setAutowired(true)
+    ->setAutowireOptionalParameters(false)
+    ->setPublic(true)
+;
+
+return $container;