+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_rot13_env.php
@@ -43,7 +43,7 @@ class Symfony_DI_PhpDumper_Test_Rot13Parameters extends Container
     public function getRemovedIds(): array
     {
         return [
-            '.service_locator.DyWBOhJ' => true,
+            '.service_locator.Qqb2hSg' => true,
         ];
     }