+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_service_locator_argument.php
@@ -44,7 +44,7 @@ class Symfony_DI_PhpDumper_Service_Locator_Argument extends Container
     public function getRemovedIds(): array
     {
         return [
-            '.service_locator.X7o4UPP' => true,
+            '.service_locator.y.HMS.V' => true,
             'foo2' => true,
             'foo3' => true,
             'foo4' => true,